Site/Site interne/requirements.txt
Flask
psycopg2-binary
Werkzeug
paho-mqtt